update actions to custom hooks

+++ src/services/apiCabins.js
@@ -1,7 +1,10 @@
 import supabase, { supabaseUrl } from "./supabase.js";
 
 export async function getCabins() {
-  const { data, error } = await supabase.from("cabins").select("*");
+  const { data, error } = await supabase
+    .from("cabins")
+    .select("*")
+    .order("created_at", { ascending: true });
 
   if (error) {
     throw new Error("Cabins could not be loaded");
@@ -10,36 +13,50 @@ export async function getCabins() {
   return data;
 }
 
-export async function createCabin(newCabin) {
+export async function createOrEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
+
   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
     "/",
     "",
   );
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
-  // 1. Create cabin
-  const { data, error } = await supabase
-    .from("cabins")
-    .insert([{ ...newCabin, image: imagePath }]);
+  // 1. Create/edit cabin
+  let query = supabase.from("cabins");
+
+  // A. Create cabin
+  if (!id) {
+    query = query.insert([{ ...newCabin, image: imagePath }]);
+  }
+
+  // B. Edit cabin
+  if (id) {
+    query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
+  }
+
+  const { data, error } = await query.select().single();
 
   if (error) {
     throw new Error("Cabin could not be created");
   }
 
   // 2. Upload image
-  const { error: storageError } = await supabase.storage
-    .from("cabin-images")
-    .upload(imageName, newCabin.image);
+  if (!hasImagePath) {
+    const { error: storageError } = await supabase.storage
+      .from("cabin-images")
+      .upload(imageName, newCabin.image);
 
-  // 3. Delete cabin if a storage error occurs when uploading image
-  if (storageError) {
-    await supabase.from("cabins").delete().eq("id", data.id);
-
-    console.log(storageError);
-
-    throw new Error(
-      "Cabin was not created because the cabin image could not be uploaded",
-    );
+    // 3. Delete cabin if a storage error occurs when uploading image
+    if (storageError) {
+      console.log(storageError);
+      await supabase.from("cabins").delete().eq("id", data.id);
+      throw new Error(
+        "Cabin was not created because the cabin image could not be uploaded",
+      );
+    }
   }
 
   return data;
